test: add loadOrRecordNockRequests method
