+++ api.planx.uk/package.json
@@ -37,6 +37,7 @@
     "dotenv": "^8.2.0",
     "graphql-query-test-mock": "^0.12.1",
     "jest": "^26.6.3",
+    "json-stringify-pretty-compact": "^3.0.0",
     "nock": "^13.0.11",
     "node-dev": "^5.0.0",
     "prettier": "^2.2.1",
